% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Anxiety_tbl_df}
\alias{Anxiety_tbl_df}
\title{Math Test Scores versus Anxiety Scores}
\format{
A tibble with 20 observations and 2 variables:
\describe{
  \item{anxiety}{Anxiety score of the student before taking the math test (integer).}
  \item{math}{Math test score of the student (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Anxiety_tbl_df)
}
\description{
This dataset, Anxiety_tbl_df, is a tibble containing information about math test scores and anxiety
levels before the test for 20 students. The dataset provides insights into the relationship between
anxiety levels and math test performance.
}
\details{
The dataset name has been kept as 'Anxiety_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
