% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Books_tbl_df}
\alias{Books_tbl_df}
\title{Books Read & Spelling Scores: 17 Third Graders}
\format{
A tibble with 17 observations and 2 variables:
\describe{
  \item{book}{Number of books read by the student (integer).}
  \item{spelling}{Spelling score of the student (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Books_tbl_df)
}
\description{
This dataset, Books_tbl_df, is a tibble containing information about the number of books read
and spelling scores for 17 third-grade students. The dataset provides insights into the relationship
between the number of books read and spelling scores in this group of students.
}
\details{
The dataset name has been kept as 'Books_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
