% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{cchousing_tbl_df}
\alias{cchousing_tbl_df}
\title{Community college housing (simulated data)}
\format{
A tibble with 75 observations and 1 variable:
\describe{
  \item{price}{Housing price for the student (numeric).}
}
}
\source{
The dataset was taken from the openintro package.
}
\usage{
data(cchousing_tbl_df)
}
\description{
This dataset, cchousing_tbl_df, is a tibble containing simulated data on housing prices for
students at a community college. The dataset provides the housing prices for 75 students,
offering insights into the distribution of housing prices in this educational setting.
}
\details{
The dataset name has been kept as 'cchousing_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
