% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{credits_tbl_df}
\alias{credits_tbl_df}
\title{College Credits}
\format{
A tibble with 100 observations and 1 variable:
\describe{
  \item{credits}{Number of college credits taken by the student (integer).}
}
}
\source{
The dataset was taken from the openintro package (simulated data).
}
\usage{
data(credits_tbl_df)
}
\description{
This dataset, credits_tbl_df, is a tibble containing simulated data on the number of college
credits taken by students each semester. The dataset includes data from 100 students, providing
insights into the distribution of credits taken by students in a college setting.
}
\details{
The dataset name has been kept as 'credits_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
