% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{crime_degree_tbl_df}
\alias{crime_degree_tbl_df}
\title{US Crime Rates vs. High School Dropout}
\format{
A tibble with 51 observations and 3 variables:
\describe{
  \item{state}{State name (character).}
  \item{nodegree}{Percentage of the population without a high school degree (numeric).}
  \item{crime}{Crime rate (numeric).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(crime_degree_tbl_df)
}
\description{
This dataset, crime_degree_tbl_df, is a tibble containing data on crime rates and the percentage
of the population without a high school degree in 51 U.S. states. The dataset includes information on
the crime rate and the percentage of the population without a high school degree for each state.
}
\details{
The dataset name has been kept as 'crime_degree_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
