% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Detroit_tbl_df}
\alias{Detroit_tbl_df}
\title{Detroit Auto Workers: Education Levels}
\format{
A tibble with 40 observations and 1 variable:
\describe{
  \item{educ}{Educational level of the auto worker (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Detroit_tbl_df)
}
\description{
This dataset, Detroit_tbl_df, is a tibble containing data on the educational levels of a sample
of 40 auto workers in Detroit. The dataset includes information on the highest level of education
attained by each worker.
}
\details{
The dataset name has been kept as 'Detroit_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
