% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Develop_tbl_df}
\alias{Develop_tbl_df}
\title{Dev Students: 2-Year & 4-Year College Demographics}
\format{
A tibble with 5656 observations and 2 variables:
\describe{
  \item{race}{Racial background of the student (factor with 5 levels).}
  \item{college}{Type of college the student is enrolled in (factor with 2 levels: 2-year or 4-year).}
}
}
\source{
The dataset was taken from "Research in Development Education" (1994), V. 11, 2.
}
\usage{
data(Develop_tbl_df)
}
\description{
This dataset, Develop_tbl_df, is a tibble containing demographic data on students enrolled
in developmental education programs at 2-year and 4-year colleges. The dataset includes
information on the racial background of the students and whether they are enrolled in
2-year or 4-year colleges.
}
\details{
The dataset name has been kept as 'Develop_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
