% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Devmath_tbl_df}
\alias{Devmath_tbl_df}
\title{Fall '95 Developmental Math: Failed Student Scores}
\format{
A tibble with 40 observations and 1 variable:
\describe{
  \item{score}{Test scores of students who failed developmental mathematics (integer).}
}
}
\source{
The dataset was taken from the BSDA package. Data provided by Dr. Anita Kitchens.
}
\usage{
data(Devmath_tbl_df)
}
\description{
This dataset, Devmath_tbl_df, is a tibble containing test scores for students who failed
developmental mathematics in the fall semester of 1995. The dataset includes the scores
of these students as part of a simulated study.
}
\details{
The dataset name has been kept as 'Devmath_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
