% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Dyslexia_tbl_df}
\alias{Dyslexia_tbl_df}
\title{Dyslexia: Data on College Students Diagnosed with Dyslexia}
\format{
A tibble with 8 observations and 7 variables:
\describe{
  \item{words}{Number of words read correctly (integer).}
  \item{age}{Age of the student (integer).}
  \item{gender}{Gender of the student (character).}
  \item{handed}{Handedness of the student (character).}
  \item{weight}{Weight of the student (integer).}
  \item{height}{Height of the student (integer).}
  \item{children}{Number of children the student has (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Dyslexia_tbl_df)
}
\description{
This dataset, Dyslexia_tbl_df, is a tibble containing data on a group of college students
diagnosed with dyslexia. The dataset includes various personal characteristics such as age,
gender, handedness, weight, height, and number of children, along with the number of words
they were able to read correctly.
}
\details{
The dataset name has been kept as 'Dyslexia_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
