% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educationR-package.R
\docType{package}
\name{educationR}
\alias{educationR}
\alias{educationR-package}
\title{educationR: A Comprehensive Collection of Educational Datasets}
\description{
{
This package provides a comprehensive collection of datasets related to education, covering topics such as student performance, learning methods, test scores, absenteeism, and other educational metrics.
}
}
\details{
educationR: A Comprehensive Collection of Educational Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of Educational Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/educationr}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Cáceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
