% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Engineer_tbl_df}
\alias{Engineer_tbl_df}
\title{Engineering Grads Salary: 10 Years Post-Graduation}
\format{
A tibble with 51 observations and 2 variables:
\describe{
  \item{salary}{Salary of the engineering graduate 10 years after graduation (integer).}
  \item{university}{Type of university the graduate attended (factor with 3 levels).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Engineer_tbl_df)
}
\description{
This dataset, Engineer_tbl_df, is a tibble containing salary data for engineering graduates
10 years after graduation. The dataset includes information on the salary of the graduates
and the type of university they graduated from (categorized into three types of universities).
}
\details{
The dataset name has been kept as 'Engineer_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
