% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Entrance_tbl_df}
\alias{Entrance_tbl_df}
\title{College entrance exam scores for 24 high school seniors}
\format{
A tibble with 24 observations and 1 variable:
\describe{
  \item{score}{College entrance exam scores (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Entrance_tbl_df)
}
\description{
This dataset, Entrance_tbl_df, is a tibble containing the college entrance exam scores
of 24 high school seniors. The dataset includes information on their exam scores.
}
\details{
The dataset name has been kept as 'Entrance_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
