% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqf_embeddings.R
\name{eqf_embeddings}
\alias{eqf_embeddings}
\title{Get word embeddings of a qualification title with respect to EQF}
\usage{
eqf_embeddings(x, locale = "en")
}
\arguments{
\item{x}{A character vector of qualification titles.}

\item{locale}{The ISO 639-1 code of the language used.}
}
\value{
Matrix with each row representing a qualification title.
}
\description{
Get word embeddings of a qualification title with respect to EQF
}
\examples{
eqf_embeddings("Bachelor of Science")
eqf_embeddings(c("Master of Science", "PhD in Linguistics"), "en")
}
