% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_eqf.R
\name{eqf_level}
\alias{eqf_level}
\title{Predict the EQF level of a qualification title}
\usage{
eqf_level(x, locale = get_languages_eqf(), na_threshold = 0.25)
}
\arguments{
\item{x}{A character vector of qualification titles.}

\item{locale}{The ISO 639-1 code of the language used. If a character vector with multiple codes is provided, the final prediction will be based on the average of predictions across those languages.}

\item{na_threshold}{Threshold for EQF level prediction score.}
}
\value{
An integer vector equal in length to x.
}
\description{
Predict the EQF level of a qualification title
}
\examples{
eqf_level("Bachelor of Science")
eqf_level(c("Master of Science", "PhD in Linguistics"), "en")
}
