% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{fheights_tbl_df}
\alias{fheights_tbl_df}
\title{Female College Student Heights}
\format{
A tibble with 24 observations and 1 variable:
\describe{
  \item{heights}{Heights of female college students, measured in inches (integer).}
}
}
\source{
The dataset was taken from the openintro package.
}
\usage{
data(fheights_tbl_df)
}
\description{
This dataset, fheights_tbl_df, is a tibble containing the heights of 24 female college students,
measured in inches. The dataset provides insight into the physical characteristics of a specific
demographic group.
}
\details{
The dataset name has been kept as 'fheights_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
