% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{German_tbl_df}
\alias{German_tbl_df}
\title{Before and after German copying errors post-course}
\format{
A tibble with 20 observations and 3 variables:
\describe{
  \item{student}{Identifier for the student (character).}
  \item{when}{Indicates whether the errors were recorded "before" or "after" the experimental course (character).}
  \item{errors}{Number of errors made by the student when copying the German passage (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(German_tbl_df)
}
\description{
This dataset, German_tbl_df, is a tibble containing data on the number of errors made by 20 students
when copying a German passage, both before and after participating in an experimental German course.
The dataset provides insights into language learning and the effectiveness of the course.
}
\details{
The dataset name has been kept as 'German_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
