% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_languages_eqf}
\alias{get_languages_eqf}
\title{List available languages for EQF prediction}
\usage{
get_languages_eqf()
}
\value{
Character vector with ISO 639-1 codes of languages supported for EQF prediction.
}
\description{
List available languages for EQF prediction
}
\examples{
get_languages_eqf()
}
