% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_languages_isced_f}
\alias{get_languages_isced_f}
\title{List available languages for ISCED-F prediction}
\usage{
get_languages_isced_f()
}
\value{
Character vector with ISO 639-1 codes of languages supported for ISCED-F prediction.
}
\description{
List available languages for ISCED-F prediction
}
\examples{
get_languages_isced_f()
}
