% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{gpa_tbl_df}
\alias{gpa_tbl_df}
\title{Survey of Duke Students on GPA, Studying, and More}
\format{
A tibble with 55 observations and 5 variables:
\describe{
  \item{gpa}{Grade Point Average (numeric).}
  \item{studyweek}{Number of hours spent studying per week (integer).}
  \item{sleepnight}{Average hours of sleep per night (numeric).}
  \item{out}{Average hours spent going out per week (numeric).}
  \item{gender}{Gender of the student (factor with levels "Male" and "Female").}
}
}
\source{
The dataset was taken from the openintro package.
}
\usage{
data(gpa_tbl_df)
}
\description{
This dataset, gpa_tbl_df, is a tibble containing survey data from 55 Duke University students.
It includes information on students' GPA, weekly study hours, average hours of sleep per night,
time spent going out per week, and gender. The dataset provides valuable insights into the
relationship between academic performance and lifestyle habits.
}
\details{
The dataset name has been kept as 'gpa_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
