% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Grades_stats_tbl_df}
\alias{Grades_stats_tbl_df}
\title{Test Grades in a Beginning Statistics Class}
\format{
A tibble with 29 observations and 1 variable:
\describe{
  \item{grades}{Test grades (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Grades_stats_tbl_df)
}
\description{
This dataset, Grades_stats_tbl_df, is a tibble containing test grades for a beginning statistics class.
It includes the grades of 29 students and provides insights into the performance distribution
in an introductory statistics course.
}
\details{
The dataset name has been kept as 'Grades_stats_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
