% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{gradestv_tbl_df}
\alias{gradestv_tbl_df}
\title{Relationship Between Watching TV and Grades (Simulated data)}
\format{
A tibble with 25 observations and 2 variables:
\describe{
  \item{tv}{Number of hours per week students watch TV (integer).}
  \item{grades}{Grade obtained in a statistics class (integer).}
}
}
\source{
The dataset was taken from the openintro package.
}
\usage{
data(gradestv_tbl_df)
}
\description{
This dataset, gradestv_tbl_df, is a tibble containing simulated data for analyzing the relationship
between the number of hours per week students watch TV and their grades in a statistics class.
It provides a simple dataset for exploring correlations or regression models in educational settings.
}
\details{
The dataset name has been kept as 'gradestv_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
