% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Graduate_tbl_df}
\alias{Graduate_tbl_df}
\title{Graduation Rates for Student Athletes in the Southeastern Conf.}
\format{
A tibble with 12 observations and 3 variables:
\describe{
  \item{school}{Name of the school (character).}
  \item{code}{Code for the school (character).}
  \item{percent}{Graduation rate as a percentage (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Graduate_tbl_df)
}
\description{
This dataset, Graduate_tbl_df, is a tibble containing data on the graduation rates for student athletes
in various schools within the Southeastern Conference. It includes the name of the school, a code for
identification, and the graduation rate as a percentage.
}
\details{
The dataset name has been kept as 'Graduate_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
