% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Habits_tbl_df}
\alias{Habits_tbl_df}
\title{Study Habits of Students in Two Matched School Districts}
\format{
A tibble with 11 observations and 4 variables:
\describe{
  \item{A}{Number of study hours in the first school district (integer).}
  \item{B}{Number of study hours in the second school district (integer).}
  \item{differ}{Difference in study hours between the two school districts (integer).}
  \item{signrks}{Significance of the difference in study hours (numeric).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Habits_tbl_df)
}
\description{
This dataset, Habits_tbl_df, is a tibble containing data on the study habits of students
in two matched school districts. It includes variables related to the number of hours students
from each district spent on studying, the difference between the two groups, and the significance
of this difference.
}
\details{
The dataset name has been kept as 'Habits_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
