% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Homework_tbl_df}
\alias{Homework_tbl_df}
\title{Homework Hours: Private vs. Public High Schools}
\format{
A tibble with 30 observations and 2 variables:
\describe{
  \item{school}{Type of school the student attends, either private or public (character).}
  \item{time}{Number of hours per week the student spends on homework (numeric).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Homework_tbl_df)
}
\description{
This dataset, Homework_tbl_df, is a tibble containing data on the number of hours
per week high school students spend on homework. The dataset compares students
from private and public schools, providing insights into the study habits and
academic workload of students in these two types of schools.
}
\details{
The dataset name has been kept as 'Homework_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
