% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Lowabil_tbl_df}
\alias{Lowabil_tbl_df}
\title{Reading skills of 24 matched low ability students}
\format{
A tibble with 12 observations and 3 variables:
\describe{
  \item{pair}{Pair identifier for the matched students (integer).}
  \item{experiment}{Reading score for students in the experimental group (integer).}
  \item{control}{Reading score for students in the control group (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Lowabil_tbl_df)
}
\description{
This dataset, Lowabil_tbl_df, is a tibble containing data on the reading skills of 24
students with low abilities, who were matched based on certain characteristics. The dataset
compares the performance of the students in two different groups: the experimental group and the control group.
This can be used to evaluate the effectiveness of an intervention or treatment in improving reading skills.
}
\details{
The dataset name has been kept as 'Lowabil_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
