% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{major_survey_tbl_df}
\alias{major_survey_tbl_df}
\title{Survey of Duke students and the area of their major}
\format{
A tibble with 218 observations and 2 variables:
\describe{
  \item{gpa}{Grade point average of the student (numeric).}
  \item{major}{Academic major of the student (factor with 3 levels).}
}
}
\source{
The dataset was taken from the openintro package.
}
\usage{
data(major_survey_tbl_df)
}
\description{
This dataset, major_survey_tbl_df, is a tibble containing information about the GPAs
of 218 Duke students and their academic major. The dataset provides insights into the
relationship between GPA and the field of study, allowing for an exploration of how
academic performance varies across different majors.
}
\details{
The dataset name has been kept as 'major_survey_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
