% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Mathcomp_tbl_df}
\alias{Mathcomp_tbl_df}
\title{Incoming Freshmen Math Competency at a Community College}
\format{
A tibble with 31 observations and 1 variable:
\describe{
  \item{score}{Standardized math competency score (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Mathcomp_tbl_df)
}
\description{
This dataset, Mathcomp_tbl_df, is a tibble containing the standardized math competency scores
for a group of 31 entering freshmen at a small community college. The dataset provides insights
into the initial math skills of incoming students, which could be useful for evaluating preparedness
and designing interventions to improve academic success.
}
\details{
The dataset name has been kept as 'Mathcomp_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
