% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Miller_scores_tbl_df}
\alias{Miller_scores_tbl_df}
\title{Graduate School Applicants' Miller Personality Test Scores}
\format{
A tibble with 25 observations and 1 variable:
\describe{
  \item{miller}{Miller Personality Test score for a student (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Miller_scores_tbl_df)
}
\description{
This dataset, Miller_scores_tbl_df, is a tibble containing the Miller Personality Test scores
for 25 college students applying for graduate school. This dataset allows for an exploration of
personality traits and their potential correlation with academic or professional success.
}
\details{
The dataset name has been kept as 'Miller_scores_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
