% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Parented_tbl_df}
\alias{Parented_tbl_df}
\title{State University Freshmen: Parental Education Profiles}
\format{
A tibble with 200 observations and 2 variables:
\describe{
  \item{education}{Education level of the parent (factor with 6 levels).}
  \item{parent}{Parent type (factor with 2 levels: "mother" or "father").}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Parented_tbl_df)
}
\description{
This dataset, Parented_tbl_df, is a tibble containing information about the education levels
of parents of 200 entering freshmen at a state university. The dataset provides insights into
the distribution of parental education backgrounds and their possible influence on students'
academic paths.
}
\details{
The dataset name has been kept as 'Parented_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
