% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_eqf.R
\name{predict_eqf}
\alias{predict_eqf}
\title{Predict the EQF level of a qualification title}
\usage{
predict_eqf(x, locale = "en")
}
\arguments{
\item{x}{A character vector of qualification titles.}

\item{locale}{The ISO 639-1 code of the language used.}
}
\value{
A factor of equal length to x with the predicted EQF levels.
}
\description{
Predict the EQF level of a qualification title
}
\examples{
predict_eqf("Bachelor of Science")
predict_eqf(c("Master of Science", "PhD in Linguistics"), "en")
}
