% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isced_f.R
\name{predict_isced_f}
\alias{predict_isced_f}
\title{Match qualification title with ISCED fields}
\usage{
predict_isced_f(x, locale = "en", target = "isced_4_label", top_docs = 5)
}
\arguments{
\item{x}{A character vector of qualification titles.}

\item{locale}{The ISO 639-1 code of the language used.}

\item{target}{Target level and formatting of output ISCED-F level.}

\item{top_docs}{Maximum number of ISCED-F fields returned per qualification title.}
}
\value{
A list of data frames with the top matched documents sorted by similarity.
}
\description{
Match qualification title with ISCED fields
}
\examples{
predict_isced_f("MSc in Biology", top_docs = 10L)
predict_isced_f(c("Law degree", "PhD in Linguistics"), "en", "isced_2_key")
}
