% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Prejudic_tbl_df}
\alias{Prejudic_tbl_df}
\title{Racial Prejudice Scores of High School Students}
\format{
A tibble with 25 observations and 1 variable:
\describe{
  \item{prejud}{Racial prejudice score (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Prejudic_tbl_df)
}
\description{
This dataset, Prejudic_tbl_df, is a tibble containing scores measuring racial prejudice
among a sample of 25 high school students. The dataset provides insights into attitudes
related to racial prejudice within this demographic.
}
\details{
The dataset name has been kept as 'Prejudic_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
