% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Program_stats_tbl_df}
\alias{Program_stats_tbl_df}
\title{Effects of Learning Methods on Statistics Students}
\format{
A tibble with 44 observations and 2 variables:
\describe{
  \item{method}{Programmed learning method (character).}
  \item{score}{Score obtained by the student (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Program_stats_tbl_df)
}
\description{
This dataset, Program_stats_tbl_df, is a tibble containing information about the
effects of four different methods of programmed learning on statistics students.
It includes the learning method used and the corresponding scores of the students.
}
\details{
The dataset name has been kept as 'Program_stats_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
