% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{PSAT_SAT_tbl_df}
\alias{PSAT_SAT_tbl_df}
\title{Relationship Between PSAT and SAT Scores}
\format{
A tibble with 7 observations and 2 variables:
\describe{
  \item{psat}{PSAT scores (integer).}
  \item{sat}{SAT scores (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(PSAT_SAT_tbl_df)
}
\description{
This dataset, PSAT_SAT_tbl_df, is a tibble containing paired data on PSAT
and SAT scores for a sample of students. It allows for the exploration of
the relationship between performance on the PSAT and SAT exams.
}
\details{
The dataset name has been kept as 'PSAT_SAT_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
