% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Readiq_tbl_df}
\alias{Readiq_tbl_df}
\title{Reading Scores vs IQ Scores}
\format{
A tibble with 14 observations and 2 variables:
\describe{
  \item{reading}{Reading scores (integer).}
  \item{iq}{IQ scores (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Readiq_tbl_df)
}
\description{
This dataset, Readiq_tbl_df, is a tibble containing paired data on reading scores
and IQ scores for a sample of individuals. It allows for the exploration of the
relationship between reading ability and IQ scores.
}
\details{
The dataset name has been kept as 'Readiq_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
