% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{sat_improve_tbl_df}
\alias{sat_improve_tbl_df}
\title{Simulated data for SAT score improvement}
\format{
A tibble with 30 observations and 1 variable:
\describe{
  \item{sat_improve}{SAT score improvement (integer).}
}
}
\source{
The dataset was taken from the openintro package.
}
\usage{
data(sat_improve_tbl_df)
}
\description{
This dataset, sat_improve_tbl_df, is a tibble containing simulated data on SAT score
improvements for a sample of students who took a course from an SAT score improvement company.
It allows for the exploration of score improvements from students who enrolled in the course.
}
\details{
The dataset name has been kept as 'sat_improve_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
