% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Schizoph_tbl_df}
\alias{Schizoph_tbl_df}
\title{Schizophrenic Learning Ability After Tranquilizer Dose}
\format{
A tibble with 13 observations and 1 variable:
\describe{
  \item{score}{Standardized exam scores (integer).}
}
}
\source{
The dataset was taken from the BSDA package.
}
\usage{
data(Schizoph_tbl_df)
}
\description{
This dataset, Schizoph_tbl_df, is a tibble containing standardized exam scores
for 13 patients. The data was collected to investigate the learning ability of
schizophrenics after being given a specified dose of a tranquilizer.
}
\details{
The dataset name has been kept as 'Schizoph_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
