% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{stats_scores_tbl_df}
\alias{stats_scores_tbl_df}
\title{Final exam scores for twenty students}
\format{
A tibble with 20 observations and 1 variable:
\describe{
  \item{scores}{Final exam scores (integer).}
}
}
\source{
The dataset was taken from the openintro package.
}
\usage{
data(stats_scores_tbl_df)
}
\description{
This dataset, stats_scores_tbl_df, is a tibble containing the final exam scores
of 20 students. The data provides insights into the academic performance of
the students in their final exam for a statistics course.
}
\details{
The dataset name has been kept as 'stats_scores_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
educationR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
