% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_eqf.R
\name{predict_eqf}
\alias{predict_eqf}
\title{Classify qualification titles with respect to EQF level}
\usage{
predict_eqf(x, locale = get_languages_eqf())
}
\arguments{
\item{x}{A character vector of qualification titles.}

\item{locale}{The ISO 639-1 code of the language used. If a character vector with multiple codes is provided, the final prediction will be the average of predictions across those languages.}
}
\value{
A matrix with as many rows as length of x and scores for each EQF level.
}
\description{
Classify qualification titles with respect to EQF level
}
\examples{
predict_eqf("Bachelor of Science")
predict_eqf(c("Master of Science", "PhD in Linguistics"), "en")
}
