% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isced_f.R
\name{predict_isced_f}
\alias{predict_isced_f}
\title{Predict the ISCED-F field of a qualification title}
\usage{
predict_isced_f(x, locale = "en", target = "isced_3_label")
}
\arguments{
\item{x}{A character vector of qualification titles.}

\item{locale}{The ISO 639-1 code of the language used.}

\item{target}{Target level and formatting of output ISCED-F level.}
}
\value{
A character vector of equal length to x with the predicted ISCED-F fields.
}
\description{
Predict the ISCED-F field of a qualification title
}
\examples{
predict_isced_f("MSc in Biology")
predict_isced_f(c("Law degree", "PhD in Linguistics"), "en", "isced_1_key")
}
