% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isced_f.R
\name{predict_isced_f}
\alias{predict_isced_f}
\title{Match qualification title with ISCED fields}
\usage{
predict_isced_f(x, locale = "en", digits = 4, top_docs = 5)
}
\arguments{
\item{x}{A character vector of qualification titles.}

\item{locale}{The ISO 639-1 code of the language used.}

\item{digits}{Number of digits of the field codes (2, 3, or 4). Determines target level of the ISCED-F hierarchy.}

\item{top_docs}{Number of top documents matched considered in matching.}
}
\value{
A list of data frames with the top matched documents sorted by similarity.
}
\description{
Match qualification title with ISCED fields
}
\examples{
predict_isced_f("MSc in Biology", top_docs = 10L)
predict_isced_f(c("Law degree", "PhD in Linguistics"), "en", 2)
}
