% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqf_embeddings.R
\name{prep_fun}
\alias{prep_fun}
\title{Preprocess character vectors}
\usage{
prep_fun(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Character vector with lower case alphanumeric text and trimmed white space.
}
\description{
Preprocess character vectors
}
